ex_1/cedric/requirements.txt
pillow == 8.1.0
numpy ==1.19.5
matplotlib~=3.3.4